Alguns elementos UI melhorados, login por Google implementado e outras coisas

+++ server/src/controllers/ImageController.ts
@@ -1,7 +1,9 @@
+import { fetchUid } from "@/firebase/fetchUid";
 import {
     createImageService,
     deleteImageService,
 } from "@/services/ImageService";
+import { UserData } from "@/utils/types/userDataType";
 import { Request, Response } from "express";
 
 /**
@@ -55,8 +57,18 @@ export const updateImage = async (
     res: Response
 ): Promise<void> => {
     try {
-        const deleteURL = req.query.url as string;
-        await deleteImageService(deleteURL);
+        if (!req.user) throw new Error("Unauthorized");
+
+        const user: UserData = await fetchUid(req.user.uid) as UserData;
+        if (!user) throw new Error("User not found");
+
+        if (
+            user.image !==
+            "https://res.cloudinary.com/di5bma0gm/image/upload/v1748032813/default_image_wroapp.png"
+        ) {
+            const deleteURL = user.image;
+            await deleteImageService(deleteURL);
+        }
 
         // Verifica se o arquivo foi enviado
         if (!req.file?.path) throw new Error("File required");
